travel_agent_system/pyproject.toml
[project]
name = "travel_agent_system"
version = "0.1.0"
description = "Multi-agent AI Travel System using CrewAI and Streamlit"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]",
    "amadeus",
    "streamlit",
    "folium",
    "streamlit-folium",
    "python-dotenv",
    "reportlab",
]

[project.scripts]
travel_agent_system = "travel_agent_system.main:main"
run_crew = "travel_agent_system.run_crew:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/travel_agent_system"]

[tool.hatch.build.targets.sdist]
packages = ["src/travel_agent_system"]
